Added battleship command, the final command to be converted from the old bot code.
